Refactor Fallow health score hotspots
