+++ apps/web/src/features/movies/catalog/mock.ts
@@ -1,0 +1,103 @@
+import type { Movie, MoviesResponse } from './types';
+
+export function buildMoviesResponse(
+  movies: Movie[],
+  page: number,
+  pageSize: number,
+): MoviesResponse {
+  const totalCount = movies.length;
+  const totalPages = Math.ceil(totalCount / pageSize);
+  const offset = (page - 1) * pageSize;
+
+  return {
+    movies: movies.slice(offset, offset + pageSize),
+    totalCount,
+    page,
+    pageSize,
+    totalPages,
+  };
+}
+
+export function generateMockMovies(): Movie[] {
+  const movies: Movie[] = [];
+  const sampleMovies = [
+    {
+      name: 'Casablanca',
+      age_rating: 'PG',
+      duration: 102,
+      score_rating: 8.5,
+      year: 1942,
+    },
+    {
+      name: 'Seven Samurai',
+      age_rating: 'NR',
+      duration: 207,
+      score_rating: 8.6,
+      year: 1954,
+    },
+    {
+      name: 'The Godfather',
+      age_rating: 'R',
+      duration: 175,
+      score_rating: 9.2,
+      year: 1972,
+    },
+    {
+      name: "One Flew Over the Cuckoo's Nest",
+      age_rating: '15',
+      duration: 133,
+      score_rating: 8.7,
+      year: 1975,
+    },
+    {
+      name: 'Star Wars: Episode IV - A New Hope',
+      age_rating: 'G',
+      duration: 121,
+      score_rating: 8.6,
+      year: 1977,
+    },
+    {
+      name: 'The Avengers',
+      age_rating: 'PG-13',
+      duration: 143,
+      score_rating: 8.0,
+      year: 2012,
+    },
+    {
+      name: "Harry Potter and the Philosopher's Stone",
+      age_rating: '12+',
+      duration: 152,
+      score_rating: 7.6,
+      year: 2001,
+    },
+    {
+      name: 'Deadpool',
+      age_rating: '16+',
+      duration: 108,
+      score_rating: 8.0,
+      year: 2016,
+    },
+    {
+      name: 'John Wick',
+      age_rating: '18+',
+      duration: 101,
+      score_rating: 7.4,
+      year: 2014,
+    },
+  ];
+
+  for (let i = 0; i < 123; i++) {
+    const baseMovie = sampleMovies[i % sampleMovies.length];
+    movies.push({
+      id: i + 1,
+      name:
+        i === 0 ? baseMovie.name : `${baseMovie.name} ${Math.floor(i / sampleMovies.length) + 1}`,
+      age_rating: baseMovie.age_rating,
+      duration: baseMovie.duration + (i % 10),
+      score_rating: Math.round((baseMovie.score_rating + (i % 20) * 0.1) * 10) / 10,
+      year: baseMovie.year + (i % 40),
+    });
+  }
+
+  return movies;
+}
