+++ apps/web/src/features/recommendation/tmdb.ts
@@ -1,733 +1,11 @@
-import z from 'zod';
-
-import { getDbClient } from '@/clients/dbClient';
-import { getOpenAIClient } from '@/clients/openaiClient';
-import {
-  extractCatalogMetadata,
-  extractUSCertification,
-  fetchMovieDetails,
-  getPosterUrl,
-} from '@/features/catalogMaintenance/tmdb';
-import { IMAGE_BASE_URL } from '@/integrations/tmdb';
-import { LOCALE_TO_TMDB_LANG } from '@/lib/locale';
-import logger from '@/lib/logger';
-import { recordOpenAIProviderError, recordTMDBProviderError } from '@/lib/metrics';
-import { MODELS } from '@/lib/models';
-import { OPENAI_TIMEOUTS_MS, openAIRequestOptions } from '@/lib/openaiTimeout';
-import { GENRE_LABEL_TO_TMDB_ID, cosineSimilarity, parseTMDBReleaseYear } from '@/lib/tmdb';
-
-import { MAX_JIT_SEED_MOVIES, MAX_TOTAL_MOVIES } from './config';
-import { formatTMDBMovieEmbeddingText, getTopTMDBGenreIds } from './tmdbDiscoverHelpers';
-
-import type { EnhancedMovieMatch, PersonFormData } from './types';
-import type { TMDBMovieDetails } from '@/features/catalogMaintenance/tmdb';
-import type { Locale } from '@/lib/locale';
-
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-
 export { MAX_TOTAL_MOVIES } from './config';
-
-/** TMDB API base URL (v3). */
-const TMDB_API_BASE = 'https://api.themoviedb.org/3';
-
-/** Timeout for TMDB discover requests in the main recommendation route. */
-const TMDB_DISCOVER_FETCH_TIMEOUT_MS = 8_000;
-const MAX_TMDB_DETAILS_ENRICHMENT = 8;
-const MAX_METADATA_QUALITY_SIMILARITY_BOOST = 0.03;
-
-const tmdbDiscoverMovieSchema = z.object({
-  id: z.number(),
-  title: z.string(),
-  overview: z.string(),
-  release_date: z.string(),
-  vote_average: z.number(),
-  vote_count: z.number().optional(),
-  popularity: z.number().optional(),
-  original_language: z.string().optional(),
-  original_title: z.string().optional(),
-  genre_ids: z.array(z.number()).optional(),
-  poster_path: z.string().nullable(),
-});
-
-export type TMDBDiscoverMovie = z.infer<typeof tmdbDiscoverMovieSchema>;
-
-const tmdbDiscoverResponseSchema = z.object({
-  results: z.array(tmdbDiscoverMovieSchema).optional(),
-});
-
-// ---------------------------------------------------------------------------
-// Internal helpers
-// ---------------------------------------------------------------------------
-
-/**
- * Derive TMDB /discover/movie query parameters from user quiz preferences.
- * Uses a deterministic mapping to avoid an extra LLM call.
- *
- * The quiz sends human-readable labels (e.g. "Sci-Fi", "New", "Serious and thought-provoking")
- * which we normalize to stable keys before mapping.
- */
-export type TMDBDiscoverQueryShape = {
-  genreIds: number[];
-  sortBy: string;
-  voteCountGte: number;
-  withoutGenreIds: number[];
-  primary_release_date_gte?: string;
-  primary_release_date_lte?: string;
-  with_runtime_lte?: number;
-};
-
-function getPreferenceText(allPeopleData: PersonFormData[]): string {
-  return allPeopleData
-    .flatMap((person) => [
-      person.favoriteMovieWhy,
-      person.newVsClassic,
-      person.tonePreference,
-      ...person.moodPreference,
-    ])
-    .filter((value): value is string => typeof value === 'string' && value.trim().length > 0)
-    .join(' ')
-    .toLocaleLowerCase('en-US');
-}
-
-function hasAnyPreferenceTerm(text: string, terms: string[]): boolean {
-  return terms.some((term) => text.includes(term));
-}
-
-export function buildTMDBDiscoverQueryShape(
-  allPeopleData: PersonFormData[],
-): TMDBDiscoverQueryShape {
-  // Top genres (up to 3) mapped to TMDB IDs
-  const preferenceText = getPreferenceText(allPeopleData);
-  const withoutGenreIds = new Set<number>();
-  if (hasAnyPreferenceTerm(preferenceText, ['avoid: horror', 'avoid horror', 'no horror'])) {
-    withoutGenreIds.add(GENRE_LABEL_TO_TMDB_ID.horror);
-  }
-  if (hasAnyPreferenceTerm(preferenceText, ['avoid: gore', 'avoid gore', 'no gore'])) {
-    withoutGenreIds.add(GENRE_LABEL_TO_TMDB_ID.horror);
-  }
-
-  const genreIds = getTopTMDBGenreIds(
-    allPeopleData.flatMap((person) => person.moodPreference),
-    { withoutGenreIds },
-  );
-
-  // Era preference: normalize to 'new' | 'classic' | 'both' by keyword matching.
-  // Quiz sends e.g. "New", "Classic", "Both new and classic".
-  const eraCounts: Record<string, number> = {};
-  allPeopleData.forEach((p) => {
-    const era = p.newVsClassic.toLowerCase();
-    let key: string;
-    if (era.includes('both')) {
-      key = 'both';
-    } else if (era.includes('classic')) {
-      key = 'classic';
-    } else if (era.includes('new')) {
-      key = 'new';
-    } else {
-      key = 'both';
-    }
-    eraCounts[key] = (eraCounts[key] ?? 0) + 1;
-  });
-  const dominantEra = Object.entries(eraCounts).sort(([, a], [, b]) => b - a)[0]?.[0] ?? 'both';
-
-  const currentYear = new Date().getFullYear();
-  let primary_release_date_gte: string | undefined;
-  let primary_release_date_lte: string | undefined;
-  if (dominantEra === 'new') {
-    primary_release_date_gte = `${currentYear - 5}-01-01`;
-  } else if (dominantEra === 'classic') {
-    primary_release_date_lte = '2000-12-31';
-  }
-
-  // Tone → sort order. Quiz sends e.g. "Serious and thought-provoking", "Dark and intense".
-  // Normalize by keyword matching.
-  const toneCounts: Record<string, number> = {};
-  allPeopleData.forEach((p) => {
-    const tone = p.tonePreference.toLowerCase();
-    let key: string;
-    if (tone.includes('serious')) {
-      key = 'serious';
-    } else if (tone.includes('dark')) {
-      key = 'dark';
-    } else if (tone.includes('balanced')) {
-      key = 'balanced';
-    } else {
-      key = 'light';
-    }
-    toneCounts[key] = (toneCounts[key] ?? 0) + 1;
-  });
-  const dominantTone = Object.entries(toneCounts).sort(([, a], [, b]) => b - a)[0]?.[0] ?? 'light';
-  let sortBy =
-    dominantTone === 'serious' || dominantTone === 'dark' ? 'vote_average.desc' : 'popularity.desc';
-
-  let voteCountGte = 100;
-  if (
-    hasAnyPreferenceTerm(preferenceText, [
-      'discovery appetite: safe hit',
-      'proven hits',
-      'familiar crowd-pleasers',
-    ])
-  ) {
-    voteCountGte = 500;
-    sortBy = 'popularity.desc';
-  } else if (
-    hasAnyPreferenceTerm(preferenceText, [
-      'discovery appetite: surprise me',
-      'unexpected picks',
-      'surprise me',
-    ])
-  ) {
-    voteCountGte = 50;
-  }
-
-  if (hasAnyPreferenceTerm(preferenceText, ['slow pacing'])) {
-    voteCountGte = Math.max(voteCountGte, 250);
-  }
-
-  const with_runtime_lte = hasAnyPreferenceTerm(preferenceText, ['long runtime']) ? 125 : undefined;
-
-  return {
-    genreIds,
-    sortBy,
-    voteCountGte,
-    withoutGenreIds: Array.from(withoutGenreIds),
-    primary_release_date_gte,
-    primary_release_date_lte,
-    with_runtime_lte,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// TMDB API
-// ---------------------------------------------------------------------------
-
-/**
- * Call TMDB /discover/movie and return up to MAX_TOTAL_MOVIES results.
- * Returns an empty array on any error so callers can treat failures gracefully.
- */
-export async function fetchTMDBDiscoverMovies(
-  allPeopleData: PersonFormData[],
-  tmdbApiKey: string,
-): Promise<TMDBDiscoverMovie[]> {
-  try {
-    const params = buildTMDBDiscoverQueryShape(allPeopleData);
-
-    const url = new URL(`${TMDB_API_BASE}/discover/movie`);
-    if (params.genreIds.length > 0) {
-      url.searchParams.set('with_genres', params.genreIds.join('|'));
-    }
-    if (params.withoutGenreIds.length > 0) {
-      url.searchParams.set('without_genres', params.withoutGenreIds.join('|'));
-    }
-    url.searchParams.set('sort_by', params.sortBy);
-    url.searchParams.set('vote_count.gte', String(params.voteCountGte));
-    url.searchParams.set('include_adult', 'false');
-    url.searchParams.set('page', '1');
-    if (params.primary_release_date_gte) {
-      url.searchParams.set('primary_release_date.gte', params.primary_release_date_gte);
-    }
-    if (params.primary_release_date_lte) {
-      url.searchParams.set('primary_release_date.lte', params.primary_release_date_lte);
-    }
-    if (params.with_runtime_lte) {
-      url.searchParams.set('with_runtime.lte', String(params.with_runtime_lte));
-    }
-
-    const response = await fetch(url.toString(), {
-      headers: {
-        Authorization: `Bearer ${tmdbApiKey}`,
-        Accept: 'application/json',
-      },
-      signal: AbortSignal.timeout(TMDB_DISCOVER_FETCH_TIMEOUT_MS),
-    });
-
-    if (!response.ok) {
-      recordTMDBProviderError('movie_discover', 'http_error');
-      logger.warn({ status: response.status }, 'TMDB discover request failed');
-      return [];
-    }
-
-    const parsedResponse = tmdbDiscoverResponseSchema.safeParse(await response.json());
-    if (!parsedResponse.success) {
-      recordTMDBProviderError('movie_discover', 'invalid_response');
-      logger.warn({ zodError: parsedResponse.error }, 'TMDB discover response validation failed');
-      return [];
-    }
-    return (parsedResponse.data.results ?? []).slice(0, MAX_TOTAL_MOVIES);
-  } catch (error) {
-    if (error instanceof Error && (error.name === 'AbortError' || error.name === 'TimeoutError')) {
-      recordTMDBProviderError('movie_discover', 'timeout');
-      logger.warn({ timeoutMs: TMDB_DISCOVER_FETCH_TIMEOUT_MS }, 'TMDB discover request timed out');
-      return [];
-    }
-    recordTMDBProviderError('movie_discover', 'error');
-    logger.warn({ err: error }, 'Error fetching movies from TMDB discover');
-    return [];
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Similarity scoring
-// ---------------------------------------------------------------------------
-
-// Re-export for backward-compatible imports from this module (e.g. route.ts).
 export { parseTMDBReleaseYear } from '@/lib/tmdb';
 
-/**
- * Embed every TMDB movie in a single batched API call, compute real cosine similarity
- * against the user's query embedding, and return fully-populated EnhancedMovieMatch objects.
- * Falls back to a neutral score (0.35) for any movie whose embedding could not be obtained.
- */
-export async function scoreAndConvertTMDBMovies(
-  movies: TMDBDiscoverMovie[],
-  queryEmbedding: number[],
-): Promise<{ matches: EnhancedMovieMatch[]; embeddings: Map<number, number[]> }> {
-  if (movies.length === 0) return { matches: [], embeddings: new Map() };
+export { buildTMDBDiscoverQueryShape, fetchTMDBDiscoverMovies } from './tmdb/discover';
+export { enrichTMDBMatchesWithDetails } from './tmdb/enrichment';
+export { scoreAndConvertTMDBMovies } from './tmdb/scoring';
+export { seedMovies, seedMoviesInBackground } from './tmdb/seeding';
+export { deserializeTMDBEmbeddings, serializeTMDBEmbeddings } from './tmdb/serialization';
 
-  const texts = movies.map(formatTMDBMovieEmbeddingText);
-
-  let rawEmbeddings: number[][] = [];
-  try {
-    const response = await getOpenAIClient().embeddings.create(
-      {
-        model: MODELS.EMBEDDING,
-        input: texts,
-      },
-      openAIRequestOptions(OPENAI_TIMEOUTS_MS.embedding),
-    );
-    rawEmbeddings = response.data.map((d) => d.embedding);
-  } catch (error) {
-    recordOpenAIProviderError('similarity_embedding', error);
-    logger.warn(
-      { err: error },
-      'Failed to embed TMDB movies for similarity scoring — using fallback score',
-    );
-  }
-
-  const embeddingsMap = new Map<number, number[]>();
-  const matches = movies.map((movie, i) => {
-    const movieEmbedding = rawEmbeddings[i];
-    if (movieEmbedding) embeddingsMap.set(movie.id, movieEmbedding);
-    const similarity = movieEmbedding
-      ? cosineSimilarity(queryEmbedding, movieEmbedding, (msg) => logger.warn(msg))
-      : 0.35;
-    return tmdbMovieToEnhancedMatch(movie, similarity);
-  });
-
-  return { matches, embeddings: embeddingsMap };
-}
-
-export async function enrichTMDBMatchesWithDetails(
-  matches: EnhancedMovieMatch[],
-  tmdbApiKey: string,
-  locale: Locale,
-): Promise<EnhancedMovieMatch[]> {
-  const tmdbMatches = matches
-    .filter((movie) => movie.tmdbId && movie.id < 0)
-    .slice(0, MAX_TMDB_DETAILS_ENRICHMENT);
-  if (tmdbMatches.length === 0) return matches;
-
-  const language = LOCALE_TO_TMDB_LANG[locale] ?? 'en-US';
-  const detailsByTMDBId = await fetchTMDBDetailsById(tmdbMatches, tmdbApiKey, language);
-  return matches.map((movie) =>
-    enrichTMDBMatchWithDetails(movie, detailsByTMDBId.get(movie.tmdbId ?? 0)),
-  );
-}
-
-async function fetchTMDBDetailsById(
-  tmdbMatches: EnhancedMovieMatch[],
-  tmdbApiKey: string,
-  language: string,
-): Promise<Map<number, TMDBMovieDetails>> {
-  const entries = await Promise.all(
-    tmdbMatches.map(async (movie) => {
-      const tmdbId = movie.tmdbId;
-      if (!tmdbId) return null;
-      const details = await fetchMovieDetails(tmdbApiKey, tmdbId, language);
-      return details ? ([tmdbId, details] as const) : null;
-    }),
-  );
-
-  return new Map(
-    entries.filter((entry): entry is readonly [number, TMDBMovieDetails] => entry !== null),
-  );
-}
-
-function markMissingTMDBDetails(movie: EnhancedMovieMatch): EnhancedMovieMatch {
-  return {
-    ...movie,
-    metadataQualityFlags: [...(movie.metadataQualityFlags ?? []), 'missing_details'],
-    metadataQualityScore: movie.metadataQualityScore ?? 0,
-  };
-}
-
-function mapTMDBWatchProviders(metadata: ReturnType<typeof extractCatalogMetadata>) {
-  return metadata.providers.map(
-    ({ availabilityType, displayPriority, providerId, providerName, region }) => ({
-      availabilityType,
-      displayPriority,
-      providerId,
-      providerName,
-      region,
-    }),
-  );
-}
-
-function shouldEnrichTMDBMatch(movie: EnhancedMovieMatch): boolean {
-  return Boolean(movie.tmdbId) && movie.id < 0;
-}
-
-function getTMDBDetailsDescription(details: TMDBMovieDetails, movie: EnhancedMovieMatch): string {
-  return details.overview || movie.description;
-}
-
-function getTMDBDetailsRuntime(details: TMDBMovieDetails, movie: EnhancedMovieMatch): number {
-  return details.runtime ?? movie.duration;
-}
-
-function getTMDBDetailsScore(details: TMDBMovieDetails, movie: EnhancedMovieMatch): number {
-  return Number((details.vote_average ?? movie.score_rating ?? 0).toFixed(1));
-}
-
-function buildTMDBDetailsContent(input: {
-  title: string;
-  releaseYear: number;
-  ageRating: string;
-  runtime: number;
-  scoreRating: number;
-  description: string;
-}): string {
-  const runtimeLabel = input.runtime ? input.runtime : 'unknown';
-  const lines = [
-    `${input.title} (${input.releaseYear}) | ${input.ageRating}`,
-    `Duration: ${runtimeLabel} min | TMDB Score: ${input.scoreRating}/10`,
-  ];
-  if (input.description) lines.push(input.description);
-  return lines.join('\n');
-}
-
-function enrichTMDBMatchWithDetails(
-  movie: EnhancedMovieMatch,
-  details?: TMDBMovieDetails,
-): EnhancedMovieMatch {
-  if (!shouldEnrichTMDBMatch(movie)) return movie;
-  if (!details) return markMissingTMDBDetails(movie);
-
-  const metadata = extractCatalogMetadata(details);
-  const ageRating = extractUSCertification(details);
-  const runtime = getTMDBDetailsRuntime(details, movie);
-  const scoreRating = getTMDBDetailsScore(details, movie);
-  const releaseYear = parseTMDBReleaseYear(details.release_date);
-  const description = getTMDBDetailsDescription(details, movie);
-  const qualityBoost = (metadata.qualityScore / 100) * MAX_METADATA_QUALITY_SIMILARITY_BOOST;
-
-  return {
-    ...movie,
-    age_rating: ageRating,
-    content: buildTMDBDetailsContent({
-      title: details.title,
-      releaseYear,
-      ageRating,
-      runtime,
-      scoreRating,
-      description,
-    }),
-    description,
-    duration: runtime,
-    metadataQualityFlags: metadata.qualityFlags,
-    metadataQualityScore: metadata.qualityScore,
-    name: details.title || movie.name,
-    originalLanguage: details.original_language ?? null,
-    popularity: details.popularity ?? null,
-    posterURL: getPosterUrl(details.poster_path) ?? movie.posterURL,
-    score_rating: scoreRating,
-    similarity: Number((movie.similarity + qualityBoost).toFixed(6)),
-    voteCount: details.vote_count ?? null,
-    watchProviders: mapTMDBWatchProviders(metadata),
-    year: releaseYear || movie.year,
-  };
-}
-
-/**
- * Convert a TMDB discover result to the EnhancedMovieMatch shape used by the rest of the route.
- * Uses a negative TMDB ID so it is distinct from positive local DB IDs.
- */
-function tmdbMovieToEnhancedMatch(
-  movie: TMDBDiscoverMovie,
-  similarity: number,
-): EnhancedMovieMatch {
-  const year = parseTMDBReleaseYear(movie.release_date);
-  const score = Number(movie.vote_average?.toFixed(1)) || 0;
-
-  const content = [`${movie.title} (${year}) | TMDB Score: ${score}/10`, movie.overview || '']
-    .filter(Boolean)
-    .join('\n');
-
-  // Carry the TMDB poster path through so enhanceSimilarMoviesWithPosters can skip its re-query
-  const posterURL = movie.poster_path ? `${IMAGE_BASE_URL}/w500${movie.poster_path}` : undefined;
-
-  return {
-    id: -movie.id, // Negative ID distinguishes TMDB-sourced movies from local DB rows (positive bigserial IDs)
-    tmdbId: movie.id,
-    name: movie.title,
-    age_rating: 'NR',
-    description: movie.overview || '',
-    duration: 0,
-    score_rating: score,
-    year,
-    similarity,
-    content,
-    originalLanguage: movie.original_language ?? null,
-    popularity: movie.popularity ?? null,
-    posterURL,
-    source: 'tmdb-discover',
-    voteCount: movie.vote_count ?? null,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// JIT seeding
-// ---------------------------------------------------------------------------
-
-/**
- * Serializable shape used when passing precomputed embeddings through queue payloads.
- */
-export type SerializableTMDBEmbeddings = Record<string, number[]>;
-
-export function serializeTMDBEmbeddings(
-  embeddings?: Map<number, number[]>,
-): SerializableTMDBEmbeddings | undefined {
-  if (!embeddings || embeddings.size === 0) return undefined;
-  return Object.fromEntries(
-    Array.from(embeddings.entries()).map(([movieId, embedding]) => [String(movieId), embedding]),
-  );
-}
-
-export function deserializeTMDBEmbeddings(
-  serialized?: SerializableTMDBEmbeddings,
-): Map<number, number[]> | undefined {
-  if (!serialized) return undefined;
-  const entries: Array<[number, number[]]> = Object.entries(serialized).flatMap(
-    ([movieId, embedding]) => {
-      const parsedMovieId = Number(movieId);
-      const isValidMovieId = Number.isFinite(parsedMovieId) && Number.isInteger(parsedMovieId);
-      const isValidEmbedding =
-        Array.isArray(embedding) &&
-        embedding.every((value) => typeof value === 'number' && Number.isFinite(value));
-
-      if (!isValidMovieId || !isValidEmbedding) return [];
-      return [[parsedMovieId, embedding]];
-    },
-  );
-  return new Map(entries);
-}
-
-function getErrorMessage(error: unknown): string {
-  if (error instanceof Error) return error.message;
-  if (typeof error === 'object' && error !== null && 'message' in error) {
-    return String((error as { message: unknown }).message);
-  }
-  return String(error);
-}
-
-function isDuplicateEntryErrorMessage(message: string): boolean {
-  const normalizedMessage = message.toLowerCase();
-  return (
-    normalizedMessage.includes('unique') ||
-    normalizedMessage.includes('duplicate') ||
-    normalizedMessage.includes('already exists')
-  );
-}
-
-function logJITSeedingError(input: { error: unknown; movieTitle: string; warnMessage: string }) {
-  if (isDuplicateEntryErrorMessage(getErrorMessage(input.error))) {
-    logger.debug(
-      { movieTitle: input.movieTitle },
-      'JIT seeding skipped — movie already in database',
-    );
-    return;
-  }
-
-  logger.warn({ err: input.error, movieTitle: input.movieTitle }, input.warnMessage);
-}
-
-function getTMDBSeedMovieKey(movie: TMDBDiscoverMovie): string {
-  return `${movie.title.toLowerCase()}|${parseTMDBReleaseYear(movie.release_date)}`;
-}
-
-function getJITSeedCandidates(
-  tmdbMovies: TMDBDiscoverMovie[],
-  existingLocalKeys: Set<string>,
-): TMDBDiscoverMovie[] {
-  return tmdbMovies
-    .filter((movie) => !existingLocalKeys.has(getTMDBSeedMovieKey(movie)))
-    .slice(0, MAX_JIT_SEED_MOVIES);
-}
-
-async function getExistingMovieKeys(
-  db: ReturnType<typeof getDbClient>,
-  candidateMovies: TMDBDiscoverMovie[],
-): Promise<Set<string>> {
-  const existingMovieKeys = new Set<string>();
-
-  try {
-    const movieNames = candidateMovies.map((movie) => movie.title);
-    const { data: existingMovies, error } = await db
-      .from<{ name: string; year: number }>('movies')
-      .select('name, year')
-      .in('name', movieNames);
-
-    if (error) {
-      logger.warn({ err: error }, 'JIT seeding existence pre-check failed');
-      return existingMovieKeys;
-    }
-
-    for (const row of existingMovies ?? []) {
-      existingMovieKeys.add(`${row.name.toLowerCase()}|${Number(row.year ?? 0)}`);
-    }
-  } catch (err) {
-    logger.warn({ err }, 'JIT seeding existence pre-check failed with unexpected error');
-  }
-
-  return existingMovieKeys;
-}
-
-async function getTMDBSeedEmbedding(
-  movie: TMDBDiscoverMovie,
-  precomputedEmbeddings?: Map<number, number[]>,
-): Promise<number[] | undefined> {
-  const precomputed = precomputedEmbeddings?.get(movie.id);
-  if (precomputed) return precomputed;
-
-  const embeddingResponse = await getOpenAIClient().embeddings.create(
-    {
-      model: MODELS.EMBEDDING,
-      input: formatTMDBMovieEmbeddingText(movie),
-    },
-    openAIRequestOptions(OPENAI_TIMEOUTS_MS.embedding),
-  );
-  return embeddingResponse.data[0]?.embedding;
-}
-
-function isZeroMagnitudeEmbedding(embedding: number[]): boolean {
-  return embedding.every((value) => value === 0);
-}
-
-async function insertTMDBSeedMovie(input: {
-  db: ReturnType<typeof getDbClient>;
-  movie: TMDBDiscoverMovie;
-  year: number;
-  score: number;
-  embedding: number[];
-}) {
-  return input.db.from('movies').insert({
-    tmdb_id: input.movie.id,
-    name: input.movie.title,
-    year: input.year,
-    age_rating: 'NR',
-    description: input.movie.overview || '',
-    duration: 0,
-    score_rating: input.score,
-    embedding: input.embedding,
-  });
-}
-
-async function seedOneTMDBMovie(input: {
-  db: ReturnType<typeof getDbClient>;
-  movie: TMDBDiscoverMovie;
-  existingMovieKeys: Set<string>;
-  precomputedEmbeddings?: Map<number, number[]>;
-}): Promise<void> {
-  const year = parseTMDBReleaseYear(input.movie.release_date);
-  const movieKey = getTMDBSeedMovieKey(input.movie);
-
-  if (input.existingMovieKeys.has(movieKey)) {
-    logger.debug(
-      { movieTitle: input.movie.title, year },
-      'JIT seeding skipped — movie already in database',
-    );
-    return;
-  }
-
-  const score = Number(input.movie.vote_average?.toFixed(1)) || 0;
-  const embedding = await getTMDBSeedEmbedding(input.movie, input.precomputedEmbeddings);
-  if (!embedding) return;
-
-  if (isZeroMagnitudeEmbedding(embedding)) {
-    logger.warn(
-      { movieTitle: input.movie.title, year },
-      'JIT seeding skipped — zero-magnitude embedding',
-    );
-    return;
-  }
-
-  const { error: insertError } = await insertTMDBSeedMovie({
-    db: input.db,
-    movie: input.movie,
-    year,
-    score,
-    embedding,
-  });
-
-  if (insertError) {
-    logJITSeedingError({
-      error: insertError,
-      movieTitle: input.movie.title,
-      warnMessage: 'JIT seeding insert failed',
-    });
-    return;
-  }
-
-  logger.info({ movieTitle: input.movie.title, year }, 'JIT seeded TMDB movie into database');
-}
-
-async function seedOneTMDBMovieWithLogging(input: {
-  db: ReturnType<typeof getDbClient>;
-  movie: TMDBDiscoverMovie;
-  existingMovieKeys: Set<string>;
-  precomputedEmbeddings?: Map<number, number[]>;
-}): Promise<void> {
-  try {
-    await seedOneTMDBMovie(input);
-  } catch (err) {
-    logJITSeedingError({
-      error: err,
-      movieTitle: input.movie.title,
-      warnMessage: 'JIT seeding failed with unexpected error',
-    });
-  }
-}
-
-/**
- * Seed TMDB movies into the local DB for future local vector search.
- */
-export async function seedMovies(
-  tmdbMovies: TMDBDiscoverMovie[],
-  existingLocalKeys: Set<string>,
-  precomputedEmbeddings?: Map<number, number[]>,
-): Promise<void> {
-  const db = getDbClient();
-  if (!db.isConfigured()) return;
-
-  const candidateMovies = getJITSeedCandidates(tmdbMovies, existingLocalKeys);
-  if (candidateMovies.length === 0) return;
-
-  const existingMovieKeys = await getExistingMovieKeys(db, candidateMovies);
-
-  for (const movie of candidateMovies) {
-    await seedOneTMDBMovieWithLogging({ db, movie, existingMovieKeys, precomputedEmbeddings });
-  }
-}
-
-/**
- * Fire-and-forget wrapper kept for fail-open fallback when queueing is unavailable.
- */
-export function seedMoviesInBackground(
-  tmdbMovies: TMDBDiscoverMovie[],
-  existingLocalKeys: Set<string>,
-  precomputedEmbeddings?: Map<number, number[]>,
-): void {
-  void seedMovies(tmdbMovies, existingLocalKeys, precomputedEmbeddings);
-}
+export type { TMDBDiscoverMovie } from './tmdb/types';
+export type { SerializableTMDBEmbeddings } from './tmdb/serialization';
