+++ apps/web/src/features/recommendation/tmdb/genreSelection.ts
@@ -1,11 +1,4 @@
-import { GENRE_LABEL_TO_TMDB_ID, normalizeGenreLabel, parseTMDBReleaseYear } from '@/lib/tmdb';
-
-export type TMDBMovieEmbeddingTextInput = {
-  title: string;
-  overview?: string | null;
-  release_date: string;
-  vote_average?: number | null;
-};
+import { GENRE_LABEL_TO_TMDB_ID, normalizeGenreLabel } from '@/lib/tmdb';
 
 function incrementCount<TKey>(counts: Map<TKey, number>, key: TKey) {
   counts.set(key, (counts.get(key) ?? 0) + 1);
@@ -23,12 +16,4 @@ export function getTopTMDBGenreIds(
     .slice(0, 3)
     .map(([key]) => GENRE_LABEL_TO_TMDB_ID[key])
     .filter((id): id is number => id !== undefined && !(options.withoutGenreIds?.has(id) ?? false));
-}
-
-export function formatTMDBMovieEmbeddingText(movie: TMDBMovieEmbeddingTextInput): string {
-  const year = parseTMDBReleaseYear(movie.release_date);
-  const score = Number(movie.vote_average?.toFixed(1)) || 0;
-  return [`${movie.title} (${year}) | TMDB Score: ${score}/10`, movie.overview || '']
-    .filter(Boolean)
-    .join('\n');
 }
