+++ docs/QUALITY.md
@@ -87,12 +87,23 @@ Completed focused follow-ups:
   and shared package health hotspots
   ([#704](https://github.com/shchilkin/PopChoice/issues/704)).
 
-Active repo-wide complexity follow-up:
+Completed repo-wide complexity follow-up:
 
 - [#717](https://github.com/shchilkin/PopChoice/issues/717): drive inherited
   Fallow complexity findings to zero actionable items. Start with the highest
   risk worker, recommendation, persistence, eval, backoffice, account/results,
   auth, catalog, and UI helper slices tracked by its child issues.
+
+Active health-score follow-up:
+
+- [#766](https://github.com/shchilkin/PopChoice/issues/766): improve the Fallow
+  health score after complexity findings reached zero. The first score-hardening
+  pass split shared DB helpers, movie catalog query handling, movie-memory
+  service helpers, recommendation TMDB modules, and the remaining coupling
+  targets. Fresh root health score after that pass: `77.8` (`B`), with penalties
+  still dominated by inherited `hotspots` and `unit_size`. Reaching the desired
+  `85-90` band requires a separate large-units wave across long production
+  pages, workers, design-system surfaces, and service/shared workflow modules.
 
 The root config intentionally ignores `**/e2e/**` for dead-code reachability and
 `collections/server` as a generated Fumadocs import. Keep those ignores narrow:
