+++ packages/shared/src/db.ts
@@ -1,1226 +1,23 @@
-import pg, { type PoolClient } from 'pg';
-
-import { CATALOG_DUPLICATE_MERGE_AUDIT_SCHEMA_SQL } from './catalogDuplicateMergeSchema.js';
-import { logger } from './logger.js';
-import { RECOMMENDATION_EVAL_RUN_SCHEMA_SQL } from './recommendationEvalSchema.js';
-
-const { Pool } = pg;
-
-export interface MovieRecord {
-  name: string;
-  year: number;
-  age_rating: string;
-  description: string;
-  duration: number;
-  score_rating: number;
-  original_title?: string | null;
-  original_language?: string | null;
-  release_date?: string | null;
-  vote_count?: number | null;
-  popularity?: number | null;
-  metadata_quality_score?: number | null;
-  metadata_quality_flags?: string[] | Record<string, unknown> | null;
-  poster_url?: string | null;
-  localized_name?: string | null;
-  tmdb_id?: number | null;
-  tmdb_match_confidence?: number | null;
-  tmdb_match_source?: 'tmdb_discovery' | 'backfill_auto' | 'manual' | null;
-  tmdb_metadata?: Record<string, unknown>;
-  tmdb_metadata_refreshed_at?: string | Date | null;
-  embedding: number[];
-}
-
-export type CatalogMetadataSource = 'tmdb' | 'manual';
-export type MoviePersonRole = 'cast' | 'director';
-
-export interface CatalogPersonRecord {
-  id: string;
-  tmdb_id: number | null;
-  name: string;
-  profile_path: string | null;
-  popularity: number | null;
-  raw_metadata: Record<string, unknown>;
-}
-
-export interface CatalogGenreRecord {
-  id: string;
-  tmdb_id: number | null;
-  name: string;
-  raw_metadata: Record<string, unknown>;
-}
-
-export interface CatalogKeywordRecord {
-  id: string;
-  tmdb_id: number | null;
-  name: string;
-  raw_metadata: Record<string, unknown>;
-}
-
-export interface MoviePersonCreditRecord {
-  id: string;
-  movie_id: string;
-  person_id: string;
-  tmdb_credit_id: string | null;
-  role: MoviePersonRole;
-  character_name: string | null;
-  job: string | null;
-  department: string | null;
-  billing_order: number | null;
-  raw_metadata: Record<string, unknown>;
-}
-
-export interface CatalogPersonInput {
-  tmdbId: number;
-  name: string;
-  profilePath?: string | null;
-  popularity?: number | null;
-  rawMetadata?: Record<string, unknown>;
-}
-
-export interface CatalogGenreInput {
-  tmdbId: number;
-  name: string;
-  rawMetadata?: Record<string, unknown>;
-}
-
-export interface CatalogKeywordInput {
-  tmdbId: number;
-  name: string;
-  rawMetadata?: Record<string, unknown>;
-}
-
-export type WatchProviderAvailabilityType = 'flatrate' | 'rent' | 'buy' | 'ads' | 'free';
-
-export interface MovieWatchProviderInput {
-  providerId: number;
-  providerName: string;
-  logoPath?: string | null;
-  displayPriority?: number | null;
-  region: string;
-  availabilityType: WatchProviderAvailabilityType;
-  link?: string | null;
-  rawMetadata?: Record<string, unknown>;
-}
-
-export interface MoviePersonCreditInput extends CatalogPersonInput {
-  creditId: string;
-  role: MoviePersonRole;
-  characterName?: string | null;
-  job?: string | null;
-  department?: string | null;
-  billingOrder?: number | null;
-}
-
-export interface MovieCatalogMetadataInput {
-  movieId: string | number;
-  tmdbMetadata?: Record<string, unknown>;
-  people?: MoviePersonCreditInput[];
-  genres?: CatalogGenreInput[];
-  keywords?: CatalogKeywordInput[];
-  providers?: MovieWatchProviderInput[];
-  source?: CatalogMetadataSource;
-}
-
-export interface InsertedMovieRecord {
-  id: string;
-  tmdb_id: number | null;
-}
-
-interface CatalogMetadataRefreshPlan {
-  movieId: string;
-  source: CatalogMetadataSource;
-  shouldRefreshPeople: boolean;
-  shouldRefreshGenres: boolean;
-  shouldRefreshKeywords: boolean;
-  shouldRefreshProviders: boolean;
-  people: MoviePersonCreditInput[];
-  genres: CatalogGenreInput[];
-  keywords: CatalogKeywordInput[];
-  providers: MovieWatchProviderInput[];
-}
-
-let pool: InstanceType<typeof Pool> | null = null;
-
-export function initDatabase(databaseUrl: string): void {
-  if (pool) return;
-  pool = new Pool({ connectionString: databaseUrl, allowExitOnIdle: true });
-}
-
-export async function closeDatabase(): Promise<void> {
-  if (!pool) return;
-  const current = pool;
-  pool = null;
-  await current.end();
-}
-
-export function getPool(): InstanceType<typeof Pool> {
-  if (!pool) throw new Error('Database pool not initialized — call initDatabase() first');
-  return pool;
-}
-
-function getMovieKey(name: string, year: number): string {
-  return `${name}\u0000${year}`;
-}
-
-function serializeMetadataQualityFlags(
-  flags: MovieRecord['metadata_quality_flags'],
-): Record<string, unknown> | string[] {
-  if (!flags) return [];
-  if (Array.isArray(flags)) return flags;
-  return flags;
-}
-
-function valueOrNull<T>(value: T | null | undefined): T | null {
-  return value ?? null;
-}
-
-function metadataJson(value: Record<string, unknown> | undefined): string {
-  return JSON.stringify(value ?? {});
-}
-
-async function getExistingMovieKeys(movies: MovieRecord[]): Promise<Set<string>> {
-  const existingKeys = new Set<string>();
-  if (movies.length === 0) return existingKeys;
-
-  const batchSize = 100;
-  for (let i = 0; i < movies.length; i += batchSize) {
-    const batch = movies.slice(i, i + batchSize);
-    const seenBatchKeys = new Set<string>();
-    const uniqueBatch: MovieRecord[] = [];
-
-    for (const movie of batch) {
-      const key = getMovieKey(movie.name, movie.year);
-      if (!seenBatchKeys.has(key)) {
-        seenBatchKeys.add(key);
-        uniqueBatch.push(movie);
-      }
-    }
-
-    const result = await getPool().query<{ name: string; year: number }>(
-      `SELECT m.name, m.year
-       FROM movies m
-       INNER JOIN unnest($1::text[], $2::int[]) AS t(n, y)
-         ON m.name = t.n AND m.year = t.y`,
-      [uniqueBatch.map((m) => m.name), uniqueBatch.map((m) => m.year)],
-    );
-
-    for (const row of result.rows) existingKeys.add(getMovieKey(row.name, row.year));
-  }
-
-  return existingKeys;
-}
-
-export async function checkTableExists(tableName: string): Promise<boolean> {
-  const result = await getPool().query(
-    'SELECT EXISTS (SELECT FROM information_schema.tables WHERE table_name = $1)',
-    [tableName],
-  );
-  return result.rows[0].exists;
-}
-
-export async function filterNewMovies(movies: MovieRecord[]): Promise<number[]> {
-  try {
-    const existingKeys = await getExistingMovieKeys(movies);
-    const newIndices: number[] = [];
-    for (let i = 0; i < movies.length; i++) {
-      const movie = movies[i];
-      if (existingKeys.has(getMovieKey(movie.name, movie.year))) {
-        logger.debug('Skipping duplicate', { name: movie.name, year: movie.year });
-      } else {
-        newIndices.push(i);
-      }
-    }
-    return newIndices;
-  } catch (err) {
-    logger.warn('Duplicate check failed, assuming all movies are new', {
-      error: err instanceof Error ? err.message : String(err),
-      movieCount: movies.length,
-    });
-    return movies.map((_, i) => i);
-  }
-}
-
-export async function ensureSchema(): Promise<void> {
-  await getPool().query('CREATE EXTENSION IF NOT EXISTS vector;');
-
-  await getPool().query(`
-    CREATE TABLE IF NOT EXISTS movies (
-      id bigserial PRIMARY KEY,
-      name text NOT NULL,
-      age_rating text NOT NULL,
-      description text NOT NULL,
-      duration integer NOT NULL,
-      score_rating float NOT NULL,
-      year int NOT NULL,
-      original_title text,
-      original_language text,
-      release_date date,
-      vote_count integer,
-      popularity float,
-      metadata_quality_score integer NOT NULL DEFAULT 0,
-      metadata_quality_flags jsonb NOT NULL DEFAULT '[]'::jsonb,
-      poster_url text,
-      localized_name text,
-      tmdb_id bigint,
-      tmdb_match_confidence float,
-      tmdb_match_source text,
-      tmdb_matched_at timestamptz,
-      embedding vector(3072),
-      UNIQUE(name, year)
-    );
-  `);
-
-  await getPool().query(`
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS poster_url text;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS localized_name text;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS tmdb_id bigint;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS tmdb_match_confidence float;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS tmdb_match_source text;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS tmdb_matched_at timestamptz;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS tmdb_metadata jsonb NOT NULL DEFAULT '{}'::jsonb;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS tmdb_metadata_refreshed_at timestamptz;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS original_title text;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS original_language text;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS release_date date;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS vote_count integer;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS popularity float;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS metadata_quality_score integer NOT NULL DEFAULT 0;
-
-    ALTER TABLE movies
-      ADD COLUMN IF NOT EXISTS metadata_quality_flags jsonb NOT NULL DEFAULT '[]'::jsonb;
-
-    ALTER TABLE movies
-      DROP CONSTRAINT IF EXISTS movies_tmdb_match_source_check;
-
-    ALTER TABLE movies
-      ADD CONSTRAINT movies_tmdb_match_source_check
-      CHECK (tmdb_match_source IS NULL OR tmdb_match_source IN ('tmdb_discovery', 'backfill_auto', 'manual'));
-
-    CREATE UNIQUE INDEX IF NOT EXISTS movies_tmdb_id_unique
-      ON movies (tmdb_id)
-      WHERE tmdb_id IS NOT NULL;
-  `);
-
-  await ensureCatalogMetadataSchema();
-
-  await getPool().query(`
-    CREATE TABLE IF NOT EXISTS tmdb_match_reviews (
-      id bigserial PRIMARY KEY,
-      movie_id bigint NOT NULL REFERENCES movies(id) ON DELETE CASCADE,
-      movie_name text NOT NULL,
-      movie_year int NOT NULL,
-      reason text NOT NULL,
-      status text NOT NULL DEFAULT 'open',
-      candidates jsonb NOT NULL DEFAULT '[]'::jsonb,
-      notes text,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT tmdb_match_reviews_reason_check CHECK (
-        reason IN ('ambiguous_match', 'runtime_mismatch')
-      ),
-      CONSTRAINT tmdb_match_reviews_status_check CHECK (
-        status IN ('open', 'resolved', 'ignored', 'deferred')
-      )
-    );
-
-    ALTER TABLE tmdb_match_reviews
-      DROP CONSTRAINT IF EXISTS tmdb_match_reviews_status_check;
-
-    ALTER TABLE tmdb_match_reviews
-      ADD CONSTRAINT tmdb_match_reviews_status_check CHECK (
-        status IN ('open', 'resolved', 'ignored', 'deferred')
-      );
-
-    CREATE UNIQUE INDEX IF NOT EXISTS idx_tmdb_match_reviews_movie_reason
-      ON tmdb_match_reviews (movie_id, reason);
-
-    CREATE INDEX IF NOT EXISTS idx_tmdb_match_reviews_status_updated_at
-      ON tmdb_match_reviews (status, updated_at DESC);
-
-    CREATE TABLE IF NOT EXISTS tmdb_match_review_audit (
-      id bigserial PRIMARY KEY,
-      review_id bigint NOT NULL REFERENCES tmdb_match_reviews(id) ON DELETE CASCADE,
-      action text NOT NULL,
-      actor text NOT NULL,
-      note text,
-      previous_status text,
-      new_status text NOT NULL,
-      candidate jsonb,
-      metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT tmdb_match_review_audit_action_check CHECK (
-        action IN ('apply_candidate', 'reject', 'defer', 'reopen')
-      ),
-      CONSTRAINT tmdb_match_review_audit_status_check CHECK (
-        new_status IN ('open', 'resolved', 'ignored', 'deferred')
-        AND (previous_status IS NULL OR previous_status IN ('open', 'resolved', 'ignored', 'deferred'))
-      )
-    );
-
-    CREATE INDEX IF NOT EXISTS idx_tmdb_match_review_audit_review_created_at
-      ON tmdb_match_review_audit (review_id, created_at DESC);
-
-    CREATE TABLE IF NOT EXISTS catalog_repair_audit (
-      id bigserial PRIMARY KEY,
-      action text NOT NULL,
-      actor text NOT NULL,
-      issue_key text NOT NULL,
-      target_type text NOT NULL,
-      target_id text NOT NULL,
-      note text,
-      previous_state jsonb NOT NULL DEFAULT '{}'::jsonb,
-      result jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT catalog_repair_audit_action_check CHECK (
-        action IN ('enqueue_backfill', 'bulk_enqueue_backfill')
-      )
-    );
-
-    ALTER TABLE catalog_repair_audit
-      DROP CONSTRAINT IF EXISTS catalog_repair_audit_action_check;
-
-    ALTER TABLE catalog_repair_audit
-      ADD CONSTRAINT catalog_repair_audit_action_check CHECK (
-        action IN ('enqueue_backfill', 'bulk_enqueue_backfill')
-      );
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_audit_target_created_at
-      ON catalog_repair_audit (target_type, target_id, created_at DESC);
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_audit_issue_created_at
-      ON catalog_repair_audit (issue_key, created_at DESC);
-
-    CREATE TABLE IF NOT EXISTS catalog_repair_batches (
-      id bigserial PRIMARY KEY,
-      action text NOT NULL,
-      actor text NOT NULL,
-      issue_key text NOT NULL,
-      target_type text NOT NULL,
-      target_id text NOT NULL,
-      status text NOT NULL DEFAULT 'enqueueing',
-      requested_limit integer NOT NULL DEFAULT 0,
-      total_candidates integer NOT NULL DEFAULT 0,
-      attempted_count integer NOT NULL DEFAULT 0,
-      queued_count integer NOT NULL DEFAULT 0,
-      deduped_count integer NOT NULL DEFAULT 0,
-      unavailable_count integer NOT NULL DEFAULT 0,
-      failed_count integer NOT NULL DEFAULT 0,
-      completed_count integer NOT NULL DEFAULT 0,
-      skipped_count integer NOT NULL DEFAULT 0,
-      note text,
-      previous_state jsonb NOT NULL DEFAULT '{}'::jsonb,
-      result jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now(),
-      completed_at timestamptz,
-      CONSTRAINT catalog_repair_batches_action_check CHECK (
-        action IN ('enqueue_backfill', 'bulk_enqueue_backfill')
-      ),
-      CONSTRAINT catalog_repair_batches_status_check CHECK (
-        status IN (
-          'enqueueing',
-          'queued',
-          'processing',
-          'partial',
-          'failed',
-          'unavailable',
-          'empty',
-          'completed'
-        )
-      )
-    );
-
-    CREATE TABLE IF NOT EXISTS catalog_repair_batch_items (
-      id bigserial PRIMARY KEY,
-      batch_id bigint NOT NULL REFERENCES catalog_repair_batches(id) ON DELETE CASCADE,
-      movie_id bigint NOT NULL REFERENCES movies(id) ON DELETE RESTRICT,
-      issue_key text NOT NULL,
-      status text NOT NULL DEFAULT 'pending',
-      queue_name text,
-      job_name text,
-      job_id text,
-      language text,
-      reason text,
-      error_message text,
-      movie_snapshot jsonb NOT NULL DEFAULT '{}'::jsonb,
-      result jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now(),
-      completed_at timestamptz,
-      CONSTRAINT catalog_repair_batch_items_status_check CHECK (
-        status IN (
-          'pending',
-          'queued',
-          'deduped',
-          'unavailable',
-          'enqueue_failed',
-          'processing',
-          'completed',
-          'completed_resolved',
-          'completed_unresolved',
-          'failed',
-          'skipped'
-        )
-      )
-    );
-
-    ALTER TABLE catalog_repair_batches
-      DROP CONSTRAINT IF EXISTS catalog_repair_batches_status_check;
-
-    ALTER TABLE catalog_repair_batches
-      ADD CONSTRAINT catalog_repair_batches_status_check CHECK (
-        status IN (
-          'enqueueing',
-          'queued',
-          'processing',
-          'partial',
-          'failed',
-          'unavailable',
-          'empty',
-          'completed'
-        )
-      );
-
-    ALTER TABLE catalog_repair_batch_items
-      DROP CONSTRAINT IF EXISTS catalog_repair_batch_items_status_check;
-
-    ALTER TABLE catalog_repair_batch_items
-      ADD CONSTRAINT catalog_repair_batch_items_status_check CHECK (
-        status IN (
-          'pending',
-          'queued',
-          'deduped',
-          'unavailable',
-          'enqueue_failed',
-          'processing',
-          'completed',
-          'completed_resolved',
-          'completed_unresolved',
-          'failed',
-          'skipped'
-        )
-      );
-
-    ALTER TABLE catalog_repair_audit
-      ADD COLUMN IF NOT EXISTS repair_batch_id bigint REFERENCES catalog_repair_batches(id) ON DELETE SET NULL;
-
-    ALTER TABLE catalog_repair_audit
-      ADD COLUMN IF NOT EXISTS repair_batch_item_id bigint REFERENCES catalog_repair_batch_items(id) ON DELETE SET NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_batches_created_at
-      ON catalog_repair_batches (created_at DESC, id DESC);
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_batches_issue_created_at
-      ON catalog_repair_batches (issue_key, created_at DESC);
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_batches_status_updated_at
-      ON catalog_repair_batches (status, updated_at DESC);
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_batch_items_batch_status
-      ON catalog_repair_batch_items (batch_id, status, id);
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_batch_items_movie_created_at
-      ON catalog_repair_batch_items (movie_id, created_at DESC);
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_batch_items_job_id
-      ON catalog_repair_batch_items (job_id)
-      WHERE job_id IS NOT NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_repair_audit_batch_created_at
-      ON catalog_repair_audit (repair_batch_id, created_at DESC)
-      WHERE repair_batch_id IS NOT NULL;
-
-    ${CATALOG_DUPLICATE_MERGE_AUDIT_SCHEMA_SQL}
-  `);
-
-  await getPool().query(RECOMMENDATION_EVAL_RUN_SCHEMA_SQL);
-
-  await getPool().query(`
-    DROP FUNCTION IF EXISTS match_movies(vector, float, int);
-
-    CREATE OR REPLACE FUNCTION match_movies (
-      query_embedding vector(3072),
-      match_threshold float DEFAULT 0.1,
-      match_count int DEFAULT 5
-    )
-    RETURNS TABLE (
-      id bigint,
-      name text,
-      age_rating text,
-      description text,
-      duration integer,
-      score_rating float,
-      year int,
-      tmdb_id bigint,
-      tmdb_match_source text,
-      original_language text,
-      vote_count integer,
-      popularity float,
-      metadata_quality_score integer,
-      metadata_quality_flags jsonb,
-      similarity float,
-      content text
-    )
-    LANGUAGE sql STABLE AS $$
-      SELECT
-        movies.id,
-        movies.name,
-        movies.age_rating,
-        movies.description,
-        movies.duration,
-        movies.score_rating,
-        movies.year,
-        movies.tmdb_id,
-        movies.tmdb_match_source,
-        movies.original_language,
-        movies.vote_count,
-        movies.popularity,
-        movies.metadata_quality_score,
-        movies.metadata_quality_flags,
-        1 - (movies.embedding <=> query_embedding) AS similarity,
-        format(
-          '%s (%s) | %s | Duration: %s min | Rating: %s/10%s%s',
-          movies.name,
-          movies.year,
-          movies.age_rating,
-          movies.duration,
-          movies.score_rating,
-          chr(10),
-          movies.description
-        ) AS content
-      FROM movies
-      WHERE movies.embedding IS NOT NULL
-        AND 1 - (movies.embedding <=> query_embedding) > match_threshold
-      ORDER BY similarity DESC
-      LIMIT match_count;
-    $$;
-  `);
-
-  logger.info('Schema ensured');
-}
-
-export async function ensureCatalogMetadataSchema(): Promise<void> {
-  await getPool().query(`
-    CREATE TABLE IF NOT EXISTS catalog_people (
-      id bigserial PRIMARY KEY,
-      tmdb_id int,
-      name text NOT NULL,
-      profile_path text,
-      popularity float,
-      raw_metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now()
-    );
-
-    ALTER TABLE catalog_people
-      ALTER COLUMN tmdb_id TYPE int USING tmdb_id::int;
-
-    CREATE UNIQUE INDEX IF NOT EXISTS catalog_people_tmdb_id_unique
-      ON catalog_people (tmdb_id)
-      WHERE tmdb_id IS NOT NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_people_name_lower
-      ON catalog_people (lower(name));
-
-    CREATE TABLE IF NOT EXISTS catalog_genres (
-      id bigserial PRIMARY KEY,
-      tmdb_id int,
-      name text NOT NULL,
-      raw_metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now()
-    );
-
-    CREATE UNIQUE INDEX IF NOT EXISTS catalog_genres_tmdb_id_unique
-      ON catalog_genres (tmdb_id)
-      WHERE tmdb_id IS NOT NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_genres_name_lower
-      ON catalog_genres (lower(name));
-
-    CREATE TABLE IF NOT EXISTS catalog_keywords (
-      id bigserial PRIMARY KEY,
-      tmdb_id int,
-      name text NOT NULL,
-      raw_metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now()
-    );
-
-    ALTER TABLE catalog_keywords
-      ALTER COLUMN tmdb_id TYPE int USING tmdb_id::int;
-
-    CREATE UNIQUE INDEX IF NOT EXISTS catalog_keywords_tmdb_id_unique
-      ON catalog_keywords (tmdb_id)
-      WHERE tmdb_id IS NOT NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_keywords_name_lower
-      ON catalog_keywords (lower(name));
-
-    CREATE TABLE IF NOT EXISTS movie_people (
-      id bigserial PRIMARY KEY,
-      movie_id bigint NOT NULL REFERENCES movies(id) ON DELETE CASCADE,
-      person_id bigint NOT NULL REFERENCES catalog_people(id) ON DELETE CASCADE,
-      tmdb_credit_id text,
-      role text NOT NULL,
-      character_name text,
-      job text,
-      department text,
-      billing_order int,
-      raw_metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT movie_people_role_check CHECK (role IN ('cast', 'director'))
-    );
-
-    CREATE UNIQUE INDEX IF NOT EXISTS movie_people_tmdb_credit_unique
-      ON movie_people (movie_id, tmdb_credit_id)
-      WHERE tmdb_credit_id IS NOT NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_movie_people_movie_role_order
-      ON movie_people (movie_id, role, billing_order NULLS LAST);
-
-    CREATE INDEX IF NOT EXISTS idx_movie_people_person_role
-      ON movie_people (person_id, role);
-
-    CREATE TABLE IF NOT EXISTS movie_genres (
-      movie_id bigint NOT NULL REFERENCES movies(id) ON DELETE CASCADE,
-      genre_id bigint NOT NULL REFERENCES catalog_genres(id) ON DELETE CASCADE,
-      source text NOT NULL DEFAULT 'tmdb',
-      created_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT movie_genres_source_check CHECK (source IN ('tmdb', 'manual')),
-      PRIMARY KEY (movie_id, genre_id)
-    );
-
-    ALTER TABLE movie_genres
-      DROP CONSTRAINT IF EXISTS movie_genres_source_check;
-
-    ALTER TABLE movie_genres
-      ADD CONSTRAINT movie_genres_source_check CHECK (source IN ('tmdb', 'manual'));
-
-    CREATE INDEX IF NOT EXISTS idx_movie_genres_genre_id
-      ON movie_genres (genre_id);
-
-    CREATE TABLE IF NOT EXISTS movie_keywords (
-      movie_id bigint NOT NULL REFERENCES movies(id) ON DELETE CASCADE,
-      keyword_id bigint NOT NULL REFERENCES catalog_keywords(id) ON DELETE CASCADE,
-      source text NOT NULL DEFAULT 'tmdb',
-      created_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT movie_keywords_source_check CHECK (source IN ('tmdb', 'manual')),
-      PRIMARY KEY (movie_id, keyword_id)
-    );
-
-    ALTER TABLE movie_keywords
-      DROP CONSTRAINT IF EXISTS movie_keywords_source_check;
-
-    ALTER TABLE movie_keywords
-      ADD CONSTRAINT movie_keywords_source_check CHECK (source IN ('tmdb', 'manual'));
-
-    CREATE INDEX IF NOT EXISTS idx_movie_keywords_keyword_id
-      ON movie_keywords (keyword_id);
-
-    CREATE TABLE IF NOT EXISTS catalog_watch_providers (
-      id bigserial PRIMARY KEY,
-      tmdb_id int,
-      provider_name text NOT NULL,
-      logo_path text,
-      raw_metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now()
-    );
-
-    CREATE UNIQUE INDEX IF NOT EXISTS catalog_watch_providers_tmdb_id_unique
-      ON catalog_watch_providers (tmdb_id)
-      WHERE tmdb_id IS NOT NULL;
-
-    CREATE INDEX IF NOT EXISTS idx_catalog_watch_providers_name_lower
-      ON catalog_watch_providers (lower(provider_name));
-
-    CREATE TABLE IF NOT EXISTS movie_watch_providers (
-      movie_id bigint NOT NULL REFERENCES movies(id) ON DELETE CASCADE,
-      provider_id bigint NOT NULL REFERENCES catalog_watch_providers(id) ON DELETE CASCADE,
-      region text NOT NULL,
-      availability_type text NOT NULL,
-      display_priority int,
-      link text,
-      raw_metadata jsonb NOT NULL DEFAULT '{}'::jsonb,
-      created_at timestamptz NOT NULL DEFAULT now(),
-      updated_at timestamptz NOT NULL DEFAULT now(),
-      CONSTRAINT movie_watch_providers_availability_type_check CHECK (
-        availability_type IN ('flatrate', 'rent', 'buy', 'ads', 'free')
-      ),
-      PRIMARY KEY (movie_id, provider_id, region, availability_type)
-    );
-
-    CREATE INDEX IF NOT EXISTS idx_movie_watch_providers_region_type
-      ON movie_watch_providers (region, availability_type);
-
-    CREATE INDEX IF NOT EXISTS idx_movie_watch_providers_provider_id
-      ON movie_watch_providers (provider_id);
-  `);
-}
-
-export async function getMovieCount(): Promise<number> {
-  const result = await getPool().query<{ count: string }>('SELECT COUNT(*) AS count FROM movies');
-  return parseInt(result.rows[0].count, 10);
-}
-
-function createCatalogMetadataRefreshPlan(
-  input: MovieCatalogMetadataInput,
-): CatalogMetadataRefreshPlan {
-  return {
-    movieId: String(input.movieId),
-    source: input.source ?? 'tmdb',
-    shouldRefreshPeople: input.people !== undefined,
-    shouldRefreshGenres: input.genres !== undefined,
-    shouldRefreshKeywords: input.keywords !== undefined,
-    shouldRefreshProviders: input.providers !== undefined,
-    people: input.people ?? [],
-    genres: input.genres ?? [],
-    keywords: input.keywords ?? [],
-    providers: input.providers ?? [],
-  };
-}
-
-async function updateMovieCatalogMetadataSnapshot(
-  client: PoolClient,
-  movieId: string,
-  tmdbMetadata: Record<string, unknown> | undefined,
-): Promise<void> {
-  if (!tmdbMetadata) return;
-
-  await client.query(
-    `UPDATE movies
-        SET tmdb_metadata = $2::jsonb,
-            tmdb_metadata_refreshed_at = now()
-      WHERE id = $1`,
-    [movieId, JSON.stringify(tmdbMetadata)],
-  );
-}
-
-async function deleteRefreshedCatalogMetadata(
-  client: PoolClient,
-  plan: CatalogMetadataRefreshPlan,
-): Promise<void> {
-  if (plan.shouldRefreshPeople) {
-    await client.query(
-      `DELETE FROM movie_people
-        WHERE movie_id = $1
-          AND tmdb_credit_id IS NOT NULL`,
-      [plan.movieId],
-    );
-  }
-  if (plan.shouldRefreshGenres) {
-    await client.query(`DELETE FROM movie_genres WHERE movie_id = $1 AND source = $2`, [
-      plan.movieId,
-      plan.source,
-    ]);
-  }
-  if (plan.shouldRefreshKeywords) {
-    await client.query(`DELETE FROM movie_keywords WHERE movie_id = $1 AND source = $2`, [
-      plan.movieId,
-      plan.source,
-    ]);
-  }
-  if (plan.shouldRefreshProviders) {
-    await client.query(`DELETE FROM movie_watch_providers WHERE movie_id = $1`, [plan.movieId]);
-  }
-}
-
-async function upsertCatalogPersonCredit(
-  client: PoolClient,
-  movieId: string,
-  person: MoviePersonCreditInput,
-): Promise<void> {
-  const personResult = await client.query<{ id: string }>(
-    `INSERT INTO catalog_people (
-       tmdb_id, name, profile_path, popularity, raw_metadata, updated_at
-     )
-     VALUES ($1, $2, $3, $4, $5::jsonb, now())
-     ON CONFLICT (tmdb_id) WHERE tmdb_id IS NOT NULL DO UPDATE
-       SET name = EXCLUDED.name,
-           profile_path = COALESCE(EXCLUDED.profile_path, catalog_people.profile_path),
-           popularity = COALESCE(EXCLUDED.popularity, catalog_people.popularity),
-           raw_metadata = EXCLUDED.raw_metadata,
-           updated_at = now()
-     RETURNING id::text`,
-    [
-      person.tmdbId,
-      person.name,
-      valueOrNull(person.profilePath),
-      valueOrNull(person.popularity),
-      metadataJson(person.rawMetadata),
-    ],
-  );
-  const personId = personResult.rows[0]?.id;
-  if (!personId) return;
-
-  await client.query(
-    `INSERT INTO movie_people (
-       movie_id, person_id, tmdb_credit_id, role, character_name, job,
-       department, billing_order, raw_metadata, updated_at
-     )
-     VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9::jsonb, now())
-     ON CONFLICT (movie_id, tmdb_credit_id) WHERE tmdb_credit_id IS NOT NULL DO UPDATE
-       SET person_id = EXCLUDED.person_id,
-           role = EXCLUDED.role,
-           character_name = EXCLUDED.character_name,
-           job = EXCLUDED.job,
-           department = EXCLUDED.department,
-           billing_order = EXCLUDED.billing_order,
-           raw_metadata = EXCLUDED.raw_metadata,
-           updated_at = now()`,
-    [
-      movieId,
-      personId,
-      person.creditId,
-      person.role,
-      valueOrNull(person.characterName),
-      valueOrNull(person.job),
-      valueOrNull(person.department),
-      valueOrNull(person.billingOrder),
-      metadataJson(person.rawMetadata),
-    ],
-  );
-}
-
-async function upsertCatalogGenre(
-  client: PoolClient,
-  movieId: string,
-  source: CatalogMetadataSource,
-  genre: CatalogGenreInput,
-): Promise<void> {
-  const genreResult = await client.query<{ id: string }>(
-    `INSERT INTO catalog_genres (tmdb_id, name, raw_metadata, updated_at)
-     VALUES ($1, $2, $3::jsonb, now())
-     ON CONFLICT (tmdb_id) WHERE tmdb_id IS NOT NULL DO UPDATE
-       SET name = EXCLUDED.name,
-           raw_metadata = EXCLUDED.raw_metadata,
-           updated_at = now()
-     RETURNING id::text`,
-    [genre.tmdbId, genre.name, metadataJson(genre.rawMetadata)],
-  );
-  const genreId = genreResult.rows[0]?.id;
-  if (!genreId) return;
-
-  await client.query(
-    `INSERT INTO movie_genres (movie_id, genre_id, source)
-     VALUES ($1, $2, $3)
-     ON CONFLICT (movie_id, genre_id) DO UPDATE
-       SET source = EXCLUDED.source`,
-    [movieId, genreId, source],
-  );
-}
-
-async function upsertCatalogKeyword(
-  client: PoolClient,
-  movieId: string,
-  source: CatalogMetadataSource,
-  keyword: CatalogKeywordInput,
-): Promise<void> {
-  const keywordResult = await client.query<{ id: string }>(
-    `INSERT INTO catalog_keywords (tmdb_id, name, raw_metadata, updated_at)
-     VALUES ($1, $2, $3::jsonb, now())
-     ON CONFLICT (tmdb_id) WHERE tmdb_id IS NOT NULL DO UPDATE
-       SET name = EXCLUDED.name,
-           raw_metadata = EXCLUDED.raw_metadata,
-           updated_at = now()
-     RETURNING id::text`,
-    [keyword.tmdbId, keyword.name, metadataJson(keyword.rawMetadata)],
-  );
-  const keywordId = keywordResult.rows[0]?.id;
-  if (!keywordId) return;
-
-  await client.query(
-    `INSERT INTO movie_keywords (movie_id, keyword_id, source)
-     VALUES ($1, $2, $3)
-     ON CONFLICT (movie_id, keyword_id) DO UPDATE
-       SET source = EXCLUDED.source`,
-    [movieId, keywordId, source],
-  );
-}
-
-async function upsertCatalogWatchProvider(
-  client: PoolClient,
-  movieId: string,
-  provider: MovieWatchProviderInput,
-): Promise<void> {
-  const providerResult = await client.query<{ id: string }>(
-    `INSERT INTO catalog_watch_providers (
-       tmdb_id, provider_name, logo_path, raw_metadata, updated_at
-     )
-     VALUES ($1, $2, $3, $4::jsonb, now())
-     ON CONFLICT (tmdb_id) WHERE tmdb_id IS NOT NULL DO UPDATE
-       SET provider_name = EXCLUDED.provider_name,
-           logo_path = COALESCE(EXCLUDED.logo_path, catalog_watch_providers.logo_path),
-           raw_metadata = EXCLUDED.raw_metadata,
-           updated_at = now()
-     RETURNING id::text`,
-    [
-      provider.providerId,
-      provider.providerName,
-      valueOrNull(provider.logoPath),
-      metadataJson(provider.rawMetadata),
-    ],
-  );
-  const providerRowId = providerResult.rows[0]?.id;
-  if (!providerRowId) return;
-
-  await client.query(
-    `INSERT INTO movie_watch_providers (
-       movie_id, provider_id, region, availability_type, display_priority,
-       link, raw_metadata, updated_at
-     )
-     VALUES ($1, $2, $3, $4, $5, $6, $7::jsonb, now())
-     ON CONFLICT (movie_id, provider_id, region, availability_type) DO UPDATE
-       SET display_priority = EXCLUDED.display_priority,
-           link = EXCLUDED.link,
-           raw_metadata = EXCLUDED.raw_metadata,
-           updated_at = now()`,
-    [
-      movieId,
-      providerRowId,
-      provider.region,
-      provider.availabilityType,
-      valueOrNull(provider.displayPriority),
-      valueOrNull(provider.link),
-      metadataJson(provider.rawMetadata),
-    ],
-  );
-}
-
-async function upsertRefreshedCatalogMetadata(
-  client: PoolClient,
-  plan: CatalogMetadataRefreshPlan,
-): Promise<void> {
-  for (const person of plan.shouldRefreshPeople ? plan.people : []) {
-    await upsertCatalogPersonCredit(client, plan.movieId, person);
-  }
-  for (const genre of plan.shouldRefreshGenres ? plan.genres : []) {
-    await upsertCatalogGenre(client, plan.movieId, plan.source, genre);
-  }
-  for (const keyword of plan.shouldRefreshKeywords ? plan.keywords : []) {
-    await upsertCatalogKeyword(client, plan.movieId, plan.source, keyword);
-  }
-  for (const provider of plan.shouldRefreshProviders ? plan.providers : []) {
-    await upsertCatalogWatchProvider(client, plan.movieId, provider);
-  }
-}
-
-export async function upsertMovieCatalogMetadata(input: MovieCatalogMetadataInput): Promise<void> {
-  const client = await getPool().connect();
-  const plan = createCatalogMetadataRefreshPlan(input);
-
-  try {
-    await client.query('BEGIN');
-    await updateMovieCatalogMetadataSnapshot(client, plan.movieId, input.tmdbMetadata);
-    await deleteRefreshedCatalogMetadata(client, plan);
-    await upsertRefreshedCatalogMetadata(client, plan);
-
-    await client.query('COMMIT');
-  } catch (error) {
-    await client.query('ROLLBACK');
-    throw error;
-  } finally {
-    client.release();
-  }
-}
-
-function normalizeInsertedMovieRows(rows: InsertedMovieRecord[]): InsertedMovieRecord[] {
-  return rows.map((row) => ({
-    id: String(row.id),
-    tmdb_id: row.tmdb_id === null ? null : Number(row.tmdb_id),
-  }));
-}
-
-function buildMovieBatchInsertParams(batch: MovieRecord[]): unknown[] {
-  return [
-    batch.map((m) => m.name),
-    batch.map((m) => m.year),
-    batch.map((m) => m.age_rating),
-    batch.map((m) => m.description),
-    batch.map((m) => m.duration),
-    batch.map((m) => m.score_rating),
-    batch.map((m) => valueOrNull(m.original_title)),
-    batch.map((m) => valueOrNull(m.original_language)),
-    batch.map((m) => valueOrNull(m.release_date)),
-    batch.map((m) => valueOrNull(m.vote_count)),
-    batch.map((m) => valueOrNull(m.popularity)),
-    batch.map((m) => m.metadata_quality_score ?? 0),
-    batch.map((m) => JSON.stringify(serializeMetadataQualityFlags(m.metadata_quality_flags))),
-    batch.map((m) => valueOrNull(m.poster_url)),
-    batch.map((m) => valueOrNull(m.localized_name)),
-    batch.map((m) => valueOrNull(m.tmdb_id)),
-    batch.map((m) => valueOrNull(m.tmdb_match_confidence)),
-    batch.map((m) => valueOrNull(m.tmdb_match_source)),
-    batch.map((m) => JSON.stringify(m.embedding)),
-  ];
-}
-
-function buildSingleMovieInsertParams(movie: MovieRecord): unknown[] {
-  return [
-    movie.name,
-    movie.year,
-    movie.age_rating,
-    movie.description,
-    movie.duration,
-    movie.score_rating,
-    valueOrNull(movie.original_title),
-    valueOrNull(movie.original_language),
-    valueOrNull(movie.release_date),
-    valueOrNull(movie.vote_count),
-    valueOrNull(movie.popularity),
-    movie.metadata_quality_score ?? 0,
-    JSON.stringify(serializeMetadataQualityFlags(movie.metadata_quality_flags)),
-    valueOrNull(movie.poster_url),
-    valueOrNull(movie.localized_name),
-    valueOrNull(movie.tmdb_id),
-    valueOrNull(movie.tmdb_match_confidence),
-    valueOrNull(movie.tmdb_match_source),
-    JSON.stringify(movie.embedding),
-  ];
-}
-
-async function insertMovieBatch(batch: MovieRecord[]): Promise<{
-  inserted: number;
-  insertedMovies: InsertedMovieRecord[];
-}> {
-  const result = await getPool().query<InsertedMovieRecord>(
-    `INSERT INTO movies (
-       name, year, age_rating, description, duration, score_rating,
-       original_title, original_language, release_date, vote_count, popularity,
-       metadata_quality_score, metadata_quality_flags, poster_url, localized_name, tmdb_id,
-       tmdb_match_confidence, tmdb_match_source, tmdb_matched_at, embedding
-     )
-     SELECT n, y, ar, d, du, sr, original, lang, release_dt, votes, pop,
-            quality_score, quality_flags::jsonb, poster, localized, tid, conf, src,
-            CASE WHEN tid IS NULL THEN NULL ELSE now() END, e::vector
-     FROM unnest(
-       $1::text[], $2::int[], $3::text[], $4::text[], $5::int[], $6::float8[],
-       $7::text[], $8::text[], $9::date[], $10::int[], $11::float8[], $12::int[],
-       $13::text[], $14::text[], $15::text[], $16::bigint[], $17::float8[],
-       $18::text[], $19::text[]
-     ) AS t(n, y, ar, d, du, sr, original, lang, release_dt, votes, pop, quality_score,
-            quality_flags, poster, localized, tid, conf, src, e)
-     ON CONFLICT (name, year) DO NOTHING
-     RETURNING id::text, tmdb_id`,
-    buildMovieBatchInsertParams(batch),
-  );
-
-  return {
-    inserted: result.rowCount ?? 0,
-    insertedMovies: normalizeInsertedMovieRows(result.rows),
-  };
-}
-
-async function insertSingleMovie(movie: MovieRecord): Promise<InsertedMovieRecord[]> {
-  const result = await getPool().query<InsertedMovieRecord>(
-    `INSERT INTO movies (
-       name, year, age_rating, description, duration, score_rating,
-       original_title, original_language, release_date, vote_count, popularity,
-       metadata_quality_score, metadata_quality_flags, poster_url, localized_name, tmdb_id,
-       tmdb_match_confidence, tmdb_match_source, tmdb_matched_at, embedding
-     )
-     VALUES (
-       $1, $2, $3, $4, $5, $6, $7::text, $8::text, $9::date, $10::int, $11::float8,
-       $12::int, $13::jsonb, $14::text, $15::text, $16::bigint,
-       $17::float8, $18::text, CASE WHEN $16 IS NULL THEN NULL ELSE now() END,
-       $19::vector
-     )
-     ON CONFLICT (name, year) DO NOTHING
-     RETURNING id::text, tmdb_id`,
-    buildSingleMovieInsertParams(movie),
-  );
-
-  return normalizeInsertedMovieRows(result.rows);
-}
-
-async function insertMovieBatchIndividually(batch: MovieRecord[]): Promise<{
-  success: number;
-  errors: number;
-  insertedMovies: InsertedMovieRecord[];
-}> {
-  let success = 0;
-  let errors = 0;
-  const insertedMovies: InsertedMovieRecord[] = [];
-
-  for (const movie of batch) {
-    try {
-      const insertedRows = await insertSingleMovie(movie);
-      success += insertedRows.length;
-      insertedMovies.push(...insertedRows);
-    } catch (singleErr) {
-      errors++;
-      logger.warn('Failed to insert movie', {
-        name: movie.name,
-        year: movie.year,
-        error: singleErr instanceof Error ? singleErr.message : String(singleErr),
-      });
-    }
-  }
-
-  return { success, errors, insertedMovies };
-}
-
-export async function insertMovies(
-  movies: MovieRecord[],
-  batchSize: number = 50,
-): Promise<{ success: number; errors: number; insertedMovies: InsertedMovieRecord[] }> {
-  let success = 0;
-  let errors = 0;
-  const insertedMovies: InsertedMovieRecord[] = [];
-
-  for (let i = 0; i < movies.length; i += batchSize) {
-    const batch = movies.slice(i, i + batchSize);
-    const batchNum = Math.floor(i / batchSize) + 1;
-
-    try {
-      const result = await insertMovieBatch(batch);
-      success += result.inserted;
-      insertedMovies.push(...result.insertedMovies);
-      logger.info('Batch inserted', {
-        batch: batchNum,
-        inserted: result.inserted,
-        total: movies.length,
-      });
-    } catch (batchErr) {
-      logger.warn('Batch insert failed, falling back to individual inserts', {
-        batch: batchNum,
-        error: batchErr instanceof Error ? batchErr.message : String(batchErr),
-      });
-      const fallback = await insertMovieBatchIndividually(batch);
-      success += fallback.success;
-      errors += fallback.errors;
-      insertedMovies.push(...fallback.insertedMovies);
-    }
-  }
-
-  return { success, errors, insertedMovies };
-}
+export { closeDatabase, getPool, initDatabase } from './db/pool.js';
+export { checkTableExists, filterNewMovies } from './db/duplicates.js';
+export { ensureCatalogMetadataSchema, ensureSchema, getMovieCount } from './db/schema.js';
+export { upsertMovieCatalogMetadata } from './db/catalogMetadata.js';
+export { insertMovies } from './db/insertMovies.js';
+
+export type {
+  CatalogGenreInput,
+  CatalogGenreRecord,
+  CatalogKeywordInput,
+  CatalogKeywordRecord,
+  CatalogMetadataSource,
+  CatalogPersonInput,
+  CatalogPersonRecord,
+  InsertedMovieRecord,
+  MovieCatalogMetadataInput,
+  MoviePersonCreditInput,
+  MoviePersonCreditRecord,
+  MoviePersonRole,
+  MovieRecord,
+  MovieWatchProviderInput,
+  WatchProviderAvailabilityType,
+} from './db/types.js';
