+++ packages/shared/src/db/pool.ts
@@ -1,0 +1,22 @@
+import pg from 'pg';
+
+const { Pool } = pg;
+
+let pool: InstanceType<typeof Pool> | null = null;
+
+export function initDatabase(databaseUrl: string): void {
+  if (pool) return;
+  pool = new Pool({ connectionString: databaseUrl, allowExitOnIdle: true });
+}
+
+export async function closeDatabase(): Promise<void> {
+  if (!pool) return;
+  const current = pool;
+  pool = null;
+  await current.end();
+}
+
+export function getPool(): InstanceType<typeof Pool> {
+  if (!pool) throw new Error('Database pool not initialized — call initDatabase() first');
+  return pool;
+}
